Changed enemy behavior and implemented explosions
Enemies now try to crash into the player. If they miss, they go past the
camera and disappear forever.
Explosions occur when enemies die or when the player is hit.
Also, input delay is reduced by using exclusive fullscreen (which will
eventually be optional) and a small change in code to take advantage of
higher refresh rates.